refactor: 로그인 페이지 코드 정리
