+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,8 +1,10 @@
-import { useState } from "react";
+import { useRef, useState } from "react";
+import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import Button from "@components/common/Button/Button";
 import LoginInput from "@components/common/logininput/loginInput";
-import { postLogin } from "@apis/LoginHttp";
+import { setUserId } from "@stores/userSlice";
+import { postauth } from "@apis/authHttp";
 import { PATH } from "@constants/path";
 import loginBackground from "@assets/loginBackground.png";
 import * as styles from "./LoginPage.style";
@@ -12,33 +14,38 @@ const LoginPage = () => {
 	const [id, setId] = useState<string>("");
 	const [pw, setPW] = useState<string>("");
 	const [errorMessage, setErrorMessage] = useState<string>("");
+	const idref = useRef<HTMLInputElement>(null);
+	const pwref = useRef<HTMLInputElement>(null);
+	const dispatch = useDispatch();
 
 	const handelLoginButtonClick = async () => {
 		setErrorMessage("");
 		// 유효성 검사
-		if (!id.trim() || !pw.trim()) {
+		if (!id.trim()) {
 			setErrorMessage("아이디와 비밀번호를 모두 입력해주세요.");
+			idref.current?.focus();
 			return;
 		}
 
-		if (id.length < 5 || id.length < 5) {
-			setErrorMessage("아이디는 5자 이상 10자 이하로 입력해주세요.");
-			return;
-		}
-
-		if (pw.length < 5 || pw.length > 16) {
-			setErrorMessage("비밀번호는 5자 이상 16자 이하로 입력해주세요.");
+		if (!pw.trim()) {
+			setErrorMessage("아이디와 비밀번호를 모두 입력해주세요.");
+			pwref.current?.focus();
 			return;
 		}
 
 		// 로그인 api 통신 처리 //브라우저에서 세션 저장 처리리
-		const result = await postLogin({ id, pw });
+		const result = await postauth({
+			id,
+			pw,
+			url: "http://localhost:8080/ureca/users/login",
+		});
 
 		console.log(result);
 
 		if (result?.status === 200 && result.data.content) {
 			console.log("서버 응답", result);
 			sessionStorage.setItem("userId", result.data.content.userId);
+			dispatch(setUserId(result?.data.content.userId));
 			nav(PATH.MAIN);
 		} else {
 			console.log("요청 실패", result);
@@ -61,12 +68,14 @@ const LoginPage = () => {
 						type="text"
 						value={id}
 						onChange={(event) => setId(event.target.value)}
+						ref={idref}
 					/>
 					<LoginInput
 						label="비밀번호"
 						type="password"
 						value={pw}
 						onChange={(event) => setPW(event.target.value)}
+						ref={pwref}
 					/>
 
 					<div className={styles.errorMessage}>{errorMessage}</div>
